Add a setting to configure the position of sub title input field.
We're using *WP Subtitle* on our site but we're rendering the sub title
string *before* the main title. To minimize confusion we want to have
the input fields in the same order. This pull request introduces a
config option which enables it to move the sub title field above the
main title field.

The setting can be found in WordPress' "Writing" settings page.

The German translation was updated with the newly introduced strings.